Side

+++ src/components/admin/channels/ChannelManagement.tsx
@@ -33,16 +33,19 @@ import DeleteIcon from "@mui/icons-material/Delete";
 import AddIcon from "@mui/icons-material/Add";
 import { z, ZodError } from "zod";
 
-import { defineAbilitiesFor } from "@/lib/abilities";
+import { AppAbility, defineAbilitiesFor } from "@/lib/abilities";
 
 import {
   createChannel,
   updateChannel,
   deleteChannel,
   fetchChannels,
+  getChannelById,
 } from "@/server-actions/channelActions";
-import { User } from "@prisma/client";
 import { UserWithRole } from "@/context/types";
+import { ChannelData } from "@/lib/typeCollection";
+import { Channel } from "@prisma/client";
+import { subject } from "@casl/ability";
 // import { useAbility } from "@casl/react";
 
 const channelSchema = z.object({
@@ -52,11 +55,11 @@ const channelSchema = z.object({
   userId: z.number().optional(),
 });
 
-type Channel = {
+type ChannelType = {
   id: number;
   name: string;
-  isActive: boolean;
-  userId?: number | null;
+  isActive: boolean | null;
+  userId: number | null;
 };
 
 interface ChannelManagementProps {
@@ -64,14 +67,14 @@ interface ChannelManagementProps {
 }
 
 const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
-  const ability = useMemo(() => defineAbilitiesFor(user), [user]);
-
   const [open, setOpen] = useState(false);
-  const [currentChannel, setCurrentChannel] = useState<Channel | null>(null);
-  const [formData, setFormData] = useState<Partial<Channel>>({});
+  const [currentChannel, setCurrentChannel] = useState<ChannelData | null>(
+    null
+  );
+  const [formData, setFormData] = useState<Partial<ChannelData>>({});
   const [validationError, setValidationError] = useState<string | null>(null);
 
-  const [channels, setChannels] = useState<Channel[]>([]);
+  const [channels, setChannels] = useState<ChannelData[]>([]);
   const [isError, setIsError] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
   const [isRefetching, setIsRefetching] = useState(false);
@@ -89,6 +92,17 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
     pageSize: 10,
   });
   console.log(user);
+
+  const [ability, setAbility] = useState<AppAbility | null>(null);
+
+  useEffect(() => {
+    const fetchAbilities = async () => {
+      const fetchedAbility = await defineAbilitiesFor(user);
+      setAbility(fetchedAbility);
+    };
+
+    fetchAbilities();
+  }, [user]);
 
   const channelsData = useCallback(async () => {
     setIsLoading(true);
@@ -124,7 +138,7 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
     channelsData();
   }, [channelsData, columnFilterFns]);
 
-  const handleOpen = (channel: Channel | null = null) => {
+  const handleOpen = (channel: ChannelData | null = null) => {
     setCurrentChannel(channel);
     setFormData(channel ? { ...channel } : {});
     setOpen(true);
@@ -148,21 +162,33 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
   console.log(formData);
   const handleSubmit = async () => {
     try {
-      const dataToSubmit = { ...formData, userId: user.id };
       channelSchema.parse(formData);
-      // setFormData();
 
-      if (currentChannel) {
-        // if (ability.can("update", currentChannel)) {
-          await updateChannel(currentChannel.id, dataToSubmit, user);
+      if (currentChannel && currentChannel.id !== undefined) {
+        const { userId } = await getChannelById(currentChannel.id);
+
+        if (
+          ability &&
+          ability.can(
+            "update",
+            subject("Channel", { ...currentChannel } as Channel)
+          )
+        ) {
+          await updateChannel(
+            currentChannel.id,
+            { ...formData, userId: userId },
+            user
+          );
           channelsData();
-        // } else {
-        //   setValidationError(
-        //     "You do not have permission to update this channel."
-        //   );
-        // }
+        } else {
+          console.log("not have permission");
+          setValidationError(
+            "You do not have permission to update this channel."
+          );
+        }
       } else {
-        if (ability.can("create", "Channel")) {
+        if (ability && ability.can("create", "Channel")) {
+          const dataToSubmit = { ...formData, userId: user.id };
           await createChannel(dataToSubmit, user);
           channelsData();
         } else {
@@ -180,20 +206,25 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
   };
 
   const handleDelete = async (id: number) => {
-    // const channelToDelete = channels.find((channel) => channel.id === id);
-    // if (
-    //   ability.can("delete", channelToDelete) &&
-    //   window.confirm("Are you sure you want to delete this channel?")
-    // ) {
+    const channelToDelete = channels.find((channel) => channel.id === id);
+    if (
+      ability &&
+      ability.can(
+        "delete",
+        subject("Channel", { ...channelToDelete } as Channel)
+      ) &&
+      window.confirm("Are you sure you want to delete this channel?")
+    ) {
       try {
         await deleteChannel(id, user);
         channelsData();
       } catch (error) {
         console.error("Error deleting channel:", error);
       }
-    // } else {
-    //   console.error("You do not have permission to delete this channel.");
-    // }
+    } else {
+      console.log("not have permission");
+      setValidationError("You do not have permission to Delete this channel.");
+    }
   };
 
   const handleColumnFiltersChange = useCallback(
@@ -246,7 +277,7 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
     ],
   };
 
-  const columns = useMemo<MRT_ColumnDef<Channel>[]>(
+  const columns = useMemo<MRT_ColumnDef<ChannelData>[]>(
     () => [
       {
         accessorKey: "id",
@@ -315,27 +346,43 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
       showProgressBars: isRefetching,
       sorting,
     },
-    renderRowActions: ({ row }) => (
-      <Box sx={{ display: "flex", gap: "1rem" }}>
-        {ability.can("update", "Channel") && (
-          <Tooltip title="Edit">
-            <IconButton onClick={() => handleOpen(row.original)}>
-              <EditIcon />
-            </IconButton>
-          </Tooltip>
-        )}
-         {ability.can("delete", "Channel") && ( 
-          <Tooltip title="Delete">
-            <IconButton
-              color="error"
-              onClick={() => handleDelete(row.original.id)}
-            >
-              <DeleteIcon />
-            </IconButton>
-          </Tooltip>
-        )} 
-      </Box>
-    ),
+    renderRowActions: ({ row }) => {
+      const ChannelToUpdate = row.original;
+      const channelToDelete = row.original;
+
+      return (
+        <Box sx={{ display: "flex", gap: "1rem" }}>
+          {ability &&
+            ability.can(
+              "update",
+              subject("Channel", ChannelToUpdate as Channel)
+            ) && (
+              <Tooltip title="Edit">
+                <IconButton onClick={() => handleOpen(row.original)}>
+                  <EditIcon />
+                </IconButton>
+              </Tooltip>
+            )}
+          {ability &&
+            ability.can(
+              "delete",
+              subject("Channel", channelToDelete as Channel)
+            ) && (
+              <Tooltip title="Delete">
+                <IconButton
+                  color="error"
+                  onClick={() =>
+                    row.original.id !== undefined &&
+                    handleDelete(row.original.id)
+                  }
+                >
+                  <DeleteIcon />
+                </IconButton>
+              </Tooltip>
+            )}
+        </Box>
+      );
+    },
 
     renderTopToolbar: () => (
       <Box
@@ -347,7 +394,7 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
           justifyContent: "space-between",
         })}
       >
-        {ability.can("create", "Channel") && ( 
+        {ability && ability.can("create", "Channel") && (
           <Button
             variant="contained"
             color="primary"
@@ -356,7 +403,7 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
           >
             Add Channel
           </Button>
-         )} 
+        )}
         <Box sx={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
           <MRT_GlobalFilterTextField table={table} />
           <MRT_ToggleFiltersButton table={table} />
@@ -381,31 +428,49 @@ const ChannelManagement: React.FC<ChannelManagementProps> = ({ user }) => {
             {currentChannel ? "Edit Channel" : "Add Channel"}
           </DialogTitle>
           <DialogContent>
-            <TextField
-              label="Channel Name"
-              name="name"
-              value={formData.name || ""}
-              onChange={handleChange}
-              fullWidth
-            />
-            {currentChannel && (
-              <Box mt={2}>
-                <FormControlLabel
-                  control={
-                    <Switch
-                      checked={formData.isActive || false}
-                      onChange={handleSwitchChange}
-                      color="primary"
+            {currentChannel ? (
+              <>
+                {ability?.can("update", "Channel", "name") && (
+                  <TextField
+                    label="Channel Name"
+                    name="name"
+                    value={formData.name || ""}
+                    onChange={handleChange}
+                    fullWidth
+                  />
+                )}
+                {ability?.can("update", "Channel", "isActive") && (
+                  <Box mt={2}>
+                    <FormControlLabel
+                      control={
+                        <Switch
+                          checked={formData.isActive || false}
+                          onChange={handleSwitchChange}
+                          color="primary"
+                        />
+                      }
+                      label="Active"
                     />
-                  }
-                  label="Active"
+                  </Box>
+                )}
+              </>
+            ) : (
+              <>
+                <TextField
+                  label="Channel Name"
+                  name="name"
+                  value={formData.name || ""}
+                  onChange={handleChange}
+                  fullWidth
                 />
-              </Box>
+              </>
             )}
+
             {validationError && (
               <p style={{ color: "red" }}>{validationError}</p>
             )}
           </DialogContent>
+
           <DialogActions>
             <Button onClick={handleClose}>Cancel</Button>
             <Button onClick={handleSubmit}>
